requirements.txt
Flask

Flask-Login
Flask-Bcrypt
Flask-SQLAlchemy
Flask-WTF
Flask-Admin

WTForms

gunicorn
pyheif
piexif
pillow
pillow-heif
numpy
scipy
werkzeug
itsdangerous
Jinja2
MarkupSafe
opencv-python 

datetime

python-dotenv
psycopg2-binary
